Make dict views behave like their unrestricted versions
unlike the restricted versions, the unrestricted versions:
 - are not iterators, they are views
 - have a len
 - are false when the mapping is empty, true otherwise
 - are instances of collections.abc.MappingView

This change aligns the behavior of restricted versions with the behavior
of unrestricted one, while keeping the "guard", which validates the
access with the security manager.

During this refactoring, also change `.items()` to validate
ach keys and values, like `.keys()` and `.values()` do.

Co-authored-by: Dieter Maurer <d-maurer@users.noreply.github.com>